+++ csvify.py
@@ -2,8 +2,8 @@
 import csv, sys, re
 
 
-for (infile,outfile) in [('parameters/beneficiaries.txt','parameters/bens.csv'),
-                         ('parameters/attributes.txt','parameters/attrs.csv')]:
+for (infile,outfile) in [('parameters/beneficiaries.txt','parameters/beneficiaries.csv'),
+                         ('parameters/attributes.txt','parameters/attributes.csv')]:
     lines = [line for line in open(infile, newline='\r\n').readlines()]
     with open(outfile, 'w') as csvfile:
         writer = csv.writer(csvfile)